Refactor of helpers -> utils
