+++ examples/sparse_templates_multi_site/main/views.py
@@ -1,13 +1,13 @@
 from django.views.generic import TemplateView
 
-from gn_django import helpers
+from gn_django import utils
 
 class Home(TemplateView):
 
     template_name = "home.j2"
 
     def get_context_data(self):
-        return {'msg': helpers.super_helper()}
+        return {'msg': utils.super_helper()}
 
 
 class Article(TemplateView):
