--- gn_django/helpers.py
@@ -1,2 +1,0 @@
-def super_helper():
-    return "DW all your problems are now fixed."
