--- gn_django/url/helpers.py
@@ -1,8 +1,0 @@
-import re
-
-def strip_protocol(url):
-    """
-    Strip protocol from URL to make it protocol relative
-    """
-    pattern = re.compile(r'^https?\:')
-    return re.sub(pattern, '', url)
